rename _registry
