+++ esp_data_temp/transforms/__init__.py
@@ -1,5 +1,5 @@
 # Has to be first because individual transform modules import register_transform
-from ._registry import register_transform, transform_from_config  # isort:skip
+from .registry import register_transform, transform_from_config  # isort:skip
 
 from .filter import Filter, FilterConfig
 from .label_from_feature import LabelFromFeature, LabelFromFeatureConfig
@@ -8,7 +8,7 @@ from .uniform_sample import UniformSample, UniformSampleConfig
 
 # It's important to export RegisteredTransformConfigs last because it's a dynamic type
 # that's updated as we import transforms.
-from ._registry import RegisteredTransformConfigs  # isort:skip
+from .registry import RegisteredTransformConfigs  # isort:skip
 
 __all__ = [
     "Filter",
